th-915: rename files
