+++ package.json
@@ -92,8 +92,8 @@
     "test:staged": "yarn test --watchAll=false --findRelatedTests",
     "test:coverage": "yarn test --coverage --watchAll=false",
     "test:debug": "react-scripts --inspect-brk test --runInBand --no-cache",
-    "browser-test": "testcafe \\\"chrome --window-size='1920,1080'\\\" testcafe/",
-    "browser-test:ci": "testcafe \\\"chrome:headless --disable-gpu --window-size='1920,1080'\\\" -s takeOnFails=true testcafe/",
+    "browser-test": "testcafe \"chrome --window-size='1920,1080'\" browser-tests/",
+    "browser-test:ci": "testcafe \"chrome:headless --disable-gpu --window-size='1920,1080'\" -s takeOnFails=true browser-tests/",
     "lint": "eslint --ext js,ts,tsx src",
     "format-code": "eslint --ext js,ts,tsx src --fix --max-warnings=0"
   },
